refactor(home): 파일명 수정 및 폴더 구조 수정
