+++ apps/home/src/App.tsx
@@ -1,5 +1,5 @@
 import "./App.css";
-import HomePage from "./pages/HomePage";
+import HomePage from "./pages/home/home-page";
 
 export default function HomeApp() {
   return <HomePage />;
