+++ apps/home/src/components/index.ts
@@ -1,5 +1,6 @@
-export { HomeHeader } from "./HomeHeader";
-export { SectionCard } from "./SectionCard";
-export { SectionGrid } from "./SectionGrid";
-export type { Section } from "../consts/sections";
+// arrow icon
+export { ArrowIcon } from "./arrow-icon";
 
+// section
+export { SectionCard } from "./section-card";
+export { SectionGrid } from "./section-grid";
