+++ apps/home/src/routes.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import type { RouteObject } from "react-router-dom";
-import HomePage from "./pages/HomePage";
+import HomePage from "./pages/home/home-page";
 
 export const routes: RouteObject = {
   children: [{ index: true, element: <HomePage /> }],
