+++ apps/home/src/types/sections.ts
@@ -1,0 +1,11 @@
+type SectionSize = "lg" | "md" | "sm";
+
+export interface Section {
+  id: string;
+  label: string;
+  description: string;
+  href: string;
+  color: string;
+  size: SectionSize;
+  tags?: string[];
+}
